update code deploy 20/10

+++ src/main/java/com/example/grocery_store_sales_online/service/impl/VariationServiceImpl.java
@@ -6,6 +6,8 @@ import com.example.grocery_store_sales_online.exception.CustomValidationExceptio
 import com.example.grocery_store_sales_online.exception.ServiceBusinessExceptional;
 import com.example.grocery_store_sales_online.custom.mapper.product.VariationMapper;
 import com.example.grocery_store_sales_online.model.product.Variation;
+import com.example.grocery_store_sales_online.model.product.VariationOption;
+import com.example.grocery_store_sales_online.projection.product.VariationProjection;
 import com.example.grocery_store_sales_online.repository.variation.IVariationJpaRepository;
 import com.example.grocery_store_sales_online.repository.variation.VariationRepository;
 import com.example.grocery_store_sales_online.service.IVariationService;
@@ -18,6 +20,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 
+import static com.example.grocery_store_sales_online.utils.CommonConstants.THIS_DATA_EDIT_FAIL;
 import static com.example.grocery_store_sales_online.utils.CommonConstants.THIS_FIELD_ALREADY_EXIST;
 
 @Service
@@ -29,10 +32,14 @@ public class VariationServiceImpl extends BaseServiceImpl implements IVariationS
     private final VariationMapper variationMapper;
 
     @Override
-    public Optional<Variation> findById(Long id) {
+    public Variation findById(Long id) {
         try {
             log.info("VariationService:findById execution started.");
-            return variationRepository.findById(id);
+            return variationRepository.findById(id)
+                    .orElseThrow(() -> new ServiceBusinessExceptional(
+                            EResponseStatus.NOT_FOUND_USER.getLabel(),
+                            EResponseStatus.NOT_FOUND_USER.getCode())
+                    );
         } catch (Exception ex) {
             log.error("Exception occurred while persisting VariationService:findById to database , Exception message {}", ex.getMessage());
             throw new ServiceBusinessExceptional(EResponseStatus.FETCH_DATA_FAIL.getLabel(), EResponseStatus.FETCH_DATA_FAIL.getCode());
@@ -51,7 +58,7 @@ public class VariationServiceImpl extends BaseServiceImpl implements IVariationS
     }
 
     @Override
-    public QueryListResult<Variation> getListResult(String queryParameter) {
+    public QueryListResult<VariationProjection> getListResult(String queryParameter) {
         try {
             log.info("VariationService:getListResult execution started.");
             return variationRepository.getListResult(readJsonQuery(queryParameter, QueryParameter.class));
@@ -87,7 +94,13 @@ public class VariationServiceImpl extends BaseServiceImpl implements IVariationS
     public void deleteModel(Long id) {
         try {
             log.info("VariationService:deleteModel execution started.");
+            if (variationRepository.findById(id).isEmpty()) {
+                throw new ServiceBusinessExceptional(EResponseStatus.NO_EXISTING.getLabel(), EResponseStatus.NO_EXISTING.getCode());
+            }
             variationRepository.deleteById(id);
+        } catch (ServiceBusinessExceptional ex) {
+            log.error("Exception occurred while  VariationService:deleteModel  , Exception message {}", ex.getMessage());
+            throw ex;
         } catch (Exception ex) {
             log.error("Exception occurred while persisting VariationService:deleteModel to database , Exception message {}", ex.getMessage());
             throw new ServiceBusinessExceptional(EResponseStatus.DELETE_FAIL.getLabel(), EResponseStatus.DELETE_FAIL.getCode());
@@ -98,10 +111,10 @@ public class VariationServiceImpl extends BaseServiceImpl implements IVariationS
     public Variation updateModelDto(Long id, VariationDto variationDto) {
         try {
             log.info("VariationService:updateModelDto execution started.");
-            Variation variation = findById(id).get();
-            if (findByName(variationDto.getName().trim()).isPresent() && !variationDto.getName().equals(variation.getName())) {
-                throw createValidationException("variationDto", "name", THIS_FIELD_ALREADY_EXIST);
-            }
+            Variation variation = variationRepository.findById(id)
+                    .orElseThrow(() ->
+                            createValidationException("variationDto", "notification", THIS_DATA_EDIT_FAIL)
+                    );
             setPersonAction(variation);
             setMetaData(variation);
             variationMapper.updateVariationFromDto(variationDto, variation);
